restructured app
