+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router, Route, Link } from "react-router-dom";
-import './index.css';
+import './css/index.css';
 import App from './App';
 import Listings from './listings';
 import * as serviceWorker from './serviceWorker';
